dmm/requirements.txt
pandas
python-amazon-sp-api
beautifulsoup4
lxml
requests